// src/app/models/model-definition-model.ts
export class ModelDefinitionModel {
  id: number = 0; //autogenerated
  modelDefCode: string = ''; //autogenerated
  modelName: string | null = null;
  assetNDId: number = 0;
  modelId: number = 0;
  countryOfOriginId: number | null = null;
  endOfSupportDate: Date | null = null;
  stopProductionDate!: Date;
  ppmFrequencyId: number | null = null;
  specialTools: string | null = null;
  modelAccessories: string | null = null;
  ppmChkLst: string | null = null;
  picName: string | null = null;
  supplierId: number | null = null;
  modelDefRelatedDefects: modelDefRelatedDefects[] = [];
  modelDefTCodes: ModelDefTCodes[] = [];
  attachments: Attachments[] = [];
  suppliers: Suppliers[] = [];
}

export class modelDefRelatedDefects {
  [x: string]: any;
  id: number = 0;
  modelDefinitionId: number = 0;
  defectId: number | null = null;
  defectName: number | null = null;
  workPerformed: string | null = null;
  cause: string | null = null;
  estimatedTime: string | null = null;
}

export class ModelDefTCodes {
  id: number = 0;
  modelDefinitionId: number = 0;
  codeTypeId: number | null = null;
  codeValue: string | null = null;
}

export class Attachments {
  id: number = 0;
  modelDefinitionId: number = 0;
  attachmentName: string | null = null;
  attachmentURL: string | null = null;
}

export class Suppliers {
  id: number = 0;
  modelDefinitionId: number = 0;
  supplierId: number | null = null;
  supplierName: string | null = null;
}

export class SearchModelDefinition
{
  pageSize:any;
  pageNumber:any;
  assetNDId:any;
  modelId:any;
  manufacturerId:any;
  constructor(){
    this.pageSize=10,
    this.pageNumber=1
  }
}